Fix handling missing hypothesmith gracefully
Fix test_mccabe not to fail to import if hypothesmith is not available.
While the original code seems to attempt to handle missing hypothesmith,
the file fails with NameError without it.  This is because hypothesis
decorators are used in global scope.  Guard the whole test function
to be defined only when hypothesmith is available.